requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
idna==3.11
numpy==2.4.1
packaging==26.0
pandas==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
setuptools==80.10.2
six==1.17.0
urllib3==2.6.3
wheel==0.46.3
